+++ text2qti/cmdline.py
@@ -29,6 +29,8 @@ def main():
     parser.add_argument('--version', action='version', version=f'text2qti {version}')
     parser.add_argument('--latex-render-url',
                         help='URL for rendering LaTeX equations')
+    parser.add_argument('--exec-code-blocks', action='store_const', const=True,
+                        help='Allow special code blocks to be executed and insert their output (off by default for security)')
     parser.add_argument('file',
                         help='File to convert from text to QTI')
     args = parser.parse_args()
@@ -55,6 +57,8 @@ def main():
             config.save()
     if args.latex_render_url is not None:
         config['latex_render_url'] = args.latex_render_url
+    if args.exec_code_blocks is not None:
+        config['exec_code_blocks'] = args.exec_code_blocks
 
     file_path = pathlib.Path(args.file)
     try:
@@ -64,6 +68,6 @@ def main():
     except UnicodeDecodeError as e:
         raise Text2qtiError(f'File "{file_path}" was not encoded in valid UTF-8:\n{e}')
 
-    quiz = Quiz(text, config=config)
+    quiz = Quiz(text, config=config, source_name=file_path.as_posix())
     qti = QTI(quiz, config=config)
     qti.save(file_path.parent / f'{file_path.stem}.zip')