+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 * Added support for essay questions.
 * Added support for question groups.
+* Added support for executable code blocks.  These can be used to generate
+  questions automatically.  This feature requires the command-line flag
+  `--exec-code-blocks` or setting `exec_code_blocks = true` in the config
+  file.
+* Fixed a bug that prevented source file names from appearing in error
+  messages.
 
 ## v0.1.0 (2020-04-01)